+++ AtlasSDK/AtlasSDKTests/APICustomerSpec.swift
@@ -45,7 +45,7 @@ class APICustomerSpec: APIClientBaseSpec {
                 let errorResponse = self.dataWithJSONObject(json)
                 let client = self.mockedAPIClient(forURL: customerUrl, data: errorResponse, statusCode: 401)
 
-                waitUntil(timeout: 10) { done in
+                waitUntil(timeout: 500) { done in
                     client.customer { result in
                         switch result {
                         case .failure(let error as AtlasAPIError):
